front/requirements.txt
Flask
google-auth
requests
